requirements.txt
fastapi
uvicorn
scikit-learn==1.4.0
pandas
requests
spacy
sentence-transformers
shap
google-auth
google-auth-oauthlib
google-api-python-client
pydantic-settings
python-dotenv
sqlalchemy
httpx
